api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.1",
    "opencv-python>=4.13.0.92",
    "pillow>=12.1.1",
    "python-multipart>=0.0.22",
    "torch>=2.10.0",
    "transformers>=5.3.0",
    "ultralytics>=8.4.21",
    "uvicorn>=0.41.0",
]
